추가적인 Navigator로 인한 충돌 문제 발견 수정 작업전 commit
