+++ .idea/misc.xml
@@ -32,6 +32,7 @@
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/activity_detail.xml" value="0.22" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/activity_main.xml" value="0.1" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/activity_my_qrcode.xml" value="0.19635416666666666" />
+        <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/activity_myinfo.xml" value="0.1" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/activity_personal.xml" value="0.2798913043478261" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/activity_terms.xml" value="0.1" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/fragment_configuration.xml" value="0.1" />
@@ -46,10 +47,11 @@
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/fragment_home_main.xml" value="0.1" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/fragment_like.xml" value="0.3359375" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/fragment_map.xml" value="0.1" />
-        <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/fragment_my_info.xml" value="0.25" />
+        <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/fragment_my_info.xml" value="0.14" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/fragment_order_list.xml" value="0.19635416666666666" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/fragment_personal.xml" value="0.1" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/fragment_restaurant_list.xml" value="0.3233695652173913" />
+        <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/fragment_sub.xml" value="0.297554347826087" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/fragment_terms.xml" value="0.1" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/notice_popup.xml" value="0.3359375" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/viewholder_cs.xml" value="0.1" />
@@ -58,6 +60,7 @@
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/viewholder_nearby_market.xml" value="0.1" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/viewholder_restaurant.xml" value="0.1" />
         <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/layout/viewholder_town_market.xml" value="0.1" />
+        <entry key="..\:/Users/HeeTae/AndroidStudioProjects/YU Market 0.2_1/app/src/main/res/menu/bottom_navigation_menu.xml" value="0.3359375" />
         <entry key="..\:/Users/StrangeCake/AndroidStudioProjects/MyApplication23/app/src/main/res/drawable/selector_navigation_color.xml" value="0.26979166666666665" />
         <entry key="..\:/Users/StrangeCake/AndroidStudioProjects/MyApplication23/app/src/main/res/layout/activity_home.xml" value="0.24003623188405798" />
         <entry key="..\:/Users/StrangeCake/AndroidStudioProjects/MyApplication23/app/src/main/res/layout/activity_main.xml" value="0.517774343122102" />
