+++ app/src/main/AndroidManifest.xml
@@ -33,9 +33,9 @@
         </activity>
         <activity android:name=".screen.myinfo.customerservice.CSActivity"/>
         <activity android:name=".screen.myinfo.customerservice.center.CSCenterActivity"/>
-        <activity android:name=".screen.myinfo.customerservice.configuration.ConfigurationActivity" />
+        <activity android:name=".screen.myinfo.customerservice.configuration.ConfigurationFragment" />
         <activity android:name=".screen.myinfo.customerservice.list.detail.CSDetailActivity"/>
-        <activity android:name=".screen.myinfo.customerservice.personal.PersonalActivity"/>
+        <activity android:name=".screen.myinfo.customerservice.personal.PersonalFragment"/>
         <activity android:name=".screen.myinfo.MyInfoActivity"/>
     </application>
