+++ app/src/main/java/com/example/myapplication23/screen/myinfo/MyInfoActivity.kt
@@ -1,9 +1,7 @@
 package com.example.myapplication23.screen.myinfo
 
 
-import android.content.Intent
 import android.os.Build
-import androidx.activity.result.ActivityResultLauncher
 import androidx.annotation.RequiresApi
 import com.example.myapplication23.R
 import com.example.myapplication23.databinding.ActivityMyinfoBinding
@@ -12,7 +10,6 @@ import com.example.myapplication23.screen.base.BaseActivity
 
 class MyInfoActivity : BaseActivity<ActivityMyinfoBinding>() {
 
-
     override fun getViewBinding(): ActivityMyinfoBinding =
         ActivityMyinfoBinding.inflate(layoutInflater)
 
@@ -20,12 +17,15 @@ class MyInfoActivity : BaseActivity<ActivityMyinfoBinding>() {
 
     }
 
+
     @RequiresApi(Build.VERSION_CODES.O)
     override fun initViews() {
         super.initViews()
 
 
        showMyInfoFragment(MyInfoFragment.newInstance(),MyInfoFragment.TAG)
+
+
     }
 
     private fun showMyInfoFragment(fragment: MyInfoFragment, tag: String) {
