+++ app/src/main/res/layout/fragment_my_info.xml
@@ -14,12 +14,20 @@
             android:layout_height="match_parent"
             android:orientation="vertical">
 
+            <!-- TODO: Update blank fragment layout -->
             <androidx.appcompat.widget.Toolbar
                 android:id="@+id/toolbar_home"
                 android:layout_width="match_parent"
                 android:layout_height="75dp"
                 app:contentInsetLeft="0dp"
                 app:contentInsetStart="0dp">
+
+                <ImageView
+                    android:id="@+id/back"
+                    android:layout_width="30dp"
+                    android:layout_height="30dp"
+                    android:src="@drawable/left"
+                   />
 
                 <RelativeLayout
                     android:id="@+id/first_layout"
@@ -77,7 +85,6 @@
 
             </androidx.appcompat.widget.Toolbar>
 
-            <!-- TODO: Update blank fragment layout -->
             <androidx.constraintlayout.widget.ConstraintLayout
                 android:layout_width="match_parent"
                 android:layout_height="match_parent">
